requirements.txt
cryptography~=43.0.0
pillow~=11.0.0
numpy~=2.1.3
